Do not bother attempting to load test app Gemfile if it doesn't exist
Helps with local development when using path: "../rails-command-connector"
in another project

+++ Gemfile
@@ -25,12 +25,16 @@ group :development, :test do
 end
 
 group :test do
-  # This loads the dependencies for the test app since it runs in memory when we run the test suite
-  test_app_gemfile_contents = File.read("spec/fixtures/rails-test-app/Gemfile")
-  test_app_gemfile_contents = test_app_gemfile_contents.gsub(/^.*foobara-rails-command-connector.*\n/, "")
-  test_app_gemfile_contents = test_app_gemfile_contents.gsub(/, path: "(.*)\/\.\.\/\.\./, ', path: "\1')
-  File.write("tmp/TestAppGemfile", test_app_gemfile_contents)
-  eval_gemfile "tmp/TestAppGemfile"
+  test_app_gemfile_path = "spec/fixtures/rails-test-app/Gemfile"
+
+  if File.exist?(test_app_gemfile_path)
+    # This loads the dependencies for the test app since it runs in memory when we run the test suite
+    test_app_gemfile_contents = File.read(test_app_gemfile_path)
+    test_app_gemfile_contents = test_app_gemfile_contents.gsub(/^.*foobara-rails-command-connector.*\n/, "")
+    test_app_gemfile_contents = test_app_gemfile_contents.gsub(/, path: "(.*)\/\.\.\/\.\./, ', path: "\1')
+    File.write("tmp/TestAppGemfile", test_app_gemfile_contents)
+    eval_gemfile "tmp/TestAppGemfile"
+  end
 
   gem "rspec-rails"
